+++ Stephen Wong CGP600 AE2/Game Files/Game/Lights/PointLight.cpp
@@ -1,7 +1,7 @@
 // *********************************************************
 //	Name:			Stephen Wong
 //	File:			PointLight.cpp
-//	Last Updated:	24/12/2017
+//	Last Updated:	03/01/2018
 //	Project:		CGP600 AE2
 // *********************************************************
 
@@ -18,7 +18,7 @@
 // Constructor
 //####################################################################################
 PointLight::PointLight(float red, float green, float blue, 
-	float x, float y, float z, int range) :
+	float x, float y, float z, float range) :
 	PositionalLight(red, green, blue, x, y, z, range)
 {
 	// Carry out positional light constructor
